requirements.txt
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
filelock==3.19.1
fonttools==4.59.1
fsspec==2025.7.0
idna==3.10
ifaddr==0.2.0
isodate==0.7.2
Jinja2==3.1.6
kiwisolver==1.4.9
lap==0.5.12
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
onvif-client==0.0.4
onvif2-zeep==0.3.4
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
pillow==11.3.0
platformdirs==4.3.8
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
requests-file==2.1.0
requests-toolbelt==1.0.0
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.184
ultralytics-thop==2.0.16
urllib3==2.5.0
WSDiscovery==2.1.2
zeep==4.3.1
